refactor(proto): move host_script and top level fetch to v1
