+++ crates/jstz_proto/src/executor/smart_function/deploy.rs
@@ -2,12 +2,12 @@ use jstz_core::{host::HostRuntime, kv::Transaction};
 use jstz_crypto::smart_function_hash::SmartFunctionHash;
 use tezos_smart_rollup::prelude::debug_msg;
 
-use crate::runtime::ParsedCode;
 use crate::{
     context::account::{Account, Addressable},
     error::Result,
     operation::DeployFunction,
     receipt::DeployFunctionReceipt,
+    runtime::ParsedCode,
     Error,
 };
 
@@ -24,7 +24,7 @@ pub fn deploy_smart_function(
     Ok(address)
 }
 
-pub(crate) fn execute(
+pub fn execute(
     hrt: &mut impl HostRuntime,
     tx: &mut Transaction,
     source: &impl Addressable,
