+++ crates/jstz_proto/src/runtime/mod.rs
@@ -1,6 +1,7 @@
 pub mod v1;
 
-pub use v1::{Kv, KvValue, LogRecord, ParsedCode, ProtocolData, LOG_PREFIX};
-
+pub use v1::{
+    run_toplevel_fetch, Kv, KvValue, LogRecord, ParsedCode, ProtocolData, LOG_PREFIX,
+};
 #[cfg(feature = "riscv")]
 pub mod v2;
